requirements.txt

discord.py[voice]
yt-dlp
python-dotenv
flask
PyNaCl
opus-tools
